backtesting/requirements.txt
requests==2.26.0
numpy==1.21.3
pandas==1.3.4
python-dotenv==0.19.1
scipy==1.7.1
jupyterlab==3.2.6
matplotlib==3.5.1
pytest==6.2.5
